+++ logprep/connector/http/input.py
@@ -27,18 +27,22 @@ Example
 
 import inspect
 import logging
+import multiprocessing as mp
 import queue
 import re
 import threading
 from abc import ABC
 from logging import Logger
-from typing import Mapping, Tuple, Union, Callable
+from typing import Callable, Mapping, Tuple, Union
 
 import falcon.asgi
 import msgspec
 import uvicorn
 from attrs import define, field, validators
-from falcon import HTTPTooManyRequests, HTTPMethodNotAllowed  # pylint: disable=no-name-in-module
+from falcon import (  # pylint: disable=no-name-in-module
+    HTTPMethodNotAllowed,
+    HTTPTooManyRequests,
+)
 
 from logprep.abc.input import FatalInputError, Input
 from logprep.util import defaults
@@ -115,7 +119,7 @@ class HttpEndpoint(ABC):
 
     Parameters
     ----------
-    messages: queue.Queue
+    messages: mp.Queue
         Input Events are put here
     collect_meta: bool
         Collects Metadata on True (default)
@@ -123,7 +127,7 @@ class HttpEndpoint(ABC):
         Defines key name for metadata
     """
 
-    def __init__(self, messages: queue.Queue, collect_meta: bool, metafield_name: str) -> None:
+    def __init__(self, messages: mp.Queue, collect_meta: bool, metafield_name: str) -> None:
         self.messages = messages
         self.collect_meta = collect_meta
         self.metafield_name = metafield_name
@@ -290,12 +294,6 @@ class ThreadingHTTPServer:  # pylint: disable=too-many-instance-attributes
 class HttpConnector(Input):
     """Connector to accept log messages as http post requests"""
 
-    _endpoint_registry: Mapping[str, HttpEndpoint] = {
-        "json": JSONHttpEndpoint,
-        "plaintext": PlaintextHttpEndpoint,
-        "jsonl": JSONLHttpEndpoint,
-    }
-
     @define(kw_only=True)
     class Config(Input.Config):
         """Config for HTTPInput"""
@@ -359,6 +357,14 @@ class HttpConnector(Input):
 
     __slots__ = []
 
+    messages: mp.Queue = None
+
+    _endpoint_registry: Mapping[str, HttpEndpoint] = {
+        "json": JSONHttpEndpoint,
+        "plaintext": PlaintextHttpEndpoint,
+        "jsonl": JSONLHttpEndpoint,
+    }
+
     def __init__(self, name: str, configuration: "HttpConnector.Config", logger: Logger) -> None:
         super().__init__(name, configuration, logger)
         internal_uvicorn_config = {
@@ -367,13 +373,9 @@ class HttpConnector(Input):
             "timeout_graceful_shutdown": 0,
         }
         self._config.uvicorn_config.update(internal_uvicorn_config)
-        self.logger = logger
         self.port = self._config.uvicorn_config["port"]
         self.host = self._config.uvicorn_config["host"]
-        self.target = "http://" + self.host + ":" + str(self.port)
-        self.messages = queue.Queue(
-            self._config.message_backlog_size
-        )  # pylint: disable=attribute-defined-outside-init
+        self.target = f"http://{self.host}:{self.port}"
 
     def setup(self):
         """setup starts the actual functionality of this connector.
@@ -386,6 +388,11 @@ class HttpConnector(Input):
             raise FatalInputError(
                 self, "Necessary instance attribute `pipeline_index` could not be found."
             )
+        self._logger.debug(
+            f"HttpInput Connector started on target {self.target} and "
+            f"queue {id(self.messages)} "
+            f"with queue_size: {self.messages._maxsize}"
+        )
         # Start HTTP Input only when in first process
         if self.pipeline_index != 1:
             return