+++ CHANGELOG.md
@@ -9,6 +9,13 @@
 
 ### Bugfix
 
+* fix a bug in http connector leading to only first process working
+
+## 11.0.1
+### Bugfix
+
+* fixes a bug where the pipeline index increases on every restart of a failed pipeline
+* fixes closed log queue issue by run logging in an extra process
 
 ## 11.0.0
 ### Breaking